Refactor notes_parser.py into a package
