+++ src/usdb_syncer/song_txt/error.py
@@ -1,0 +1,5 @@
+"""Errors when parsing an UltraStar txt file."""
+
+
+class NotesParseError(Exception):
+    """Raised when failing to parse notes."""
